+++ src/main/java/com/todoseventos/todos_eventos/model/cliente/ClienteJuridicaModel.java
@@ -13,7 +13,7 @@ import java.io.Serializable;
 @Builder(toBuilder = true)
 public class ClienteJuridicaModel implements Serializable {
 
-    private Long idPessoa;
+    private Integer idPessoa;
     private String cnpj;
     private String nome;
     private String email;
